commit_0413_1700
We have also updated the code to include option
to run in parallel. This is not required for the logit model estimation, but may be required for the nested logit model estimation,
or for calculating elasticities.

We run with 2000 agents, and using the input data
in which we drop products with less than 50 sales for a given model
year.